tools/requirements.txt
pydicom>=2.3.0
pillow>=9.0.0
numpy>=1.21.0